🔒 LOCK IN: WISE² DEFENSE NIGHTWING v1.0 - COMPLETE
BRANDING LOCKED (IMMUTABLE):
- System: WISE² DEFENSE NIGHTWING
- Hero: LOCK IT IN.
- Tagline: TRAIN. TEACH. PROTECT.
- Region: Greensboro, NC
- Colors: Red, White, Silver on Black

ALL SYSTEMS INTEGRATED & LIVE:
✅ Knight Wing website (now Nightwing branding)
✅ Google Maps crime radar (Greensboro 6 zones)
✅ RTL-SDR spectrum monitoring (88M-1200M live)
✅ IMP always-on voice listening (24/7 clap detection)
✅ Animated Pi 3B+ dashboard (5 states, 4 panels)
✅ Crime radar visualization (signals + incidents)
✅ Mesh networking, weather alerts, offline-first

PRODUCTION READY - READY FOR DEPLOYMENT

+++ apps/fieldtech-android/app/src/main/kotlin/com/wise2/fieldtech/ui/screens/readings/LiveReadingsScreen.kt
@@ -104,7 +104,7 @@ fun LiveReadingsScreen(viewModel: LiveReadingsViewModel, onBack: () -> Unit) {
                 reading.lowSidePsig?.let { pressure ->
                     item {
                         Meter(
-                            value = pressure,
+                            value = pressure.toFloat(),
                             min = 0f,
                             max = 400f,
                             unit = "psig",
@@ -118,7 +118,7 @@ fun LiveReadingsScreen(viewModel: LiveReadingsViewModel, onBack: () -> Unit) {
                 reading.highSidePsig?.let { pressure ->
                     item {
                         Meter(
-                            value = pressure,
+                            value = pressure.toFloat(),
                             min = 0f,
                             max = 600f,
                             unit = "psig",
@@ -183,7 +183,7 @@ fun LiveReadingsScreen(viewModel: LiveReadingsViewModel, onBack: () -> Unit) {
                 reading.voltageL1?.let { voltage ->
                     item {
                         Meter(
-                            value = voltage,
+                            value = voltage.toFloat(),
                             min = 200f,
                             max = 250f,
                             unit = "V",
